Expose meta in Route objects
Expose route.meta to allow reading RouteConfig.meta in transition hooks.

Reasons: 
In 1.0, we used to pass Custom Fields( like `auth` in route configs, as the [document](http://router.vuejs.org/en/route.html#custom-fields) suggested), but in 2.0 these custom fields aren't exposed in `route` object.
I've noticed that in `RouteConfig`, there is an optional field `meta`, which i guess serves the functionality of the previous Custom Fields, judging from the fact that 1). It's type is any 2).there are no references of it in the source code. 
So I guess `meta` should be exposed in route objects, otherwise there is no need to define it in the first place.