feat(core): publish 3 new properties on $route.currentRoute (children, parent, title)
It is not easy to get access to child routes when you need to render nested menus with <router-link>
Publishing "children" on the $route.currentRoute solves a real problem. The "parent" property is required so that the route can see its siblings. The "title" property is optional - it is useful instead of "name" when you want to name a route having a default child. I personally use "name" in beforeRouteEnter/beforeRouteUpdate to update the "document.title" in browser - and I need it on all routes (even with default child), so I had to add "title" property.

This closes issue #1149