add this.$router information
Include information about the this.$router syntax. For many Vue.js projects vue-router will be configured in the main.js file and programmatic routes will be called from component files. This syntax is incredibly helpful and not obvious in the documentation. I found out about it because of [this thread](https://laracasts.com/discuss/channels/vue/redirect-from-method-in-vuejs-with-vue-router). Figured it should be accessible for all! Thank you