Prevent exceptions when a consumer calls `isSameRoute` with falsey `a`
In certain cases, I've encountered an exception where `isSameRoute` is
called by an upstream consumer where a is undefined. This is despite the
fact that the type declaration explicitely prohibits such a thing.

Unfortunately, due to my own incompetence at reading the way the vue
error handling stack works; I've been unable to isolate the *caller*.

So, I've added a test case where `a` is a falsey value; in the hope that
this will prevent an exception from being raised, which was triggering
an infinite loop in my particular case; which, again unfortunately, I couldn't figure
out higher up the stack.

Since this is possibly a bug in the calling library, I'm not really
sure it's worth fixing here; but if it is fixed here it'd save me a few
more hours or days of debugging.